[refactor] all the classes earlier in odemis.gui.comp.overlay.view.py have been split into individual python files.
