+++ src/odemis/gui/comp/overlay/view_select.py
@@ -1,0 +1,91 @@
+# -*- coding: utf-8 -*-
+
+"""
+:created: 2014-01-25
+:author: Rinze de Laat
+:copyright: © 2014 Rinze de Laat, Delmic
+
+This file is part of Odemis.
+
+.. license::
+    Odemis is free software: you can redistribute it and/or modify it under the
+    terms of the GNU General Public License version 2 as published by the Free
+    Software Foundation.
+
+    Odemis is distributed in the hope that it will be useful, but WITHOUT ANY
+    WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
+    PARTICULAR PURPOSE. See the GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License along with
+    Odemis. If not, see http://www.gnu.org/licenses/.
+
+"""
+
+import cairo
+
+import odemis.gui as gui
+import odemis.gui.comp.overlay.base as base
+
+
+class ViewSelectOverlay(base.ViewOverlay, base.SelectionMixin):
+
+    def __init__(self, cnvs, colour=gui.SELECTION_COLOUR, center=(0, 0)):
+        base.ViewOverlay.__init__(self, cnvs)
+        base.SelectionMixin.__init__(self, colour, center, base.EDIT_MODE_BOX)
+
+        self.position_label = self.add_label("")
+
+    def draw(self, ctx, shift=(0, 0), scale=1.0):
+
+        if self.select_v_start_pos and self.select_v_end_pos:
+            start_pos = self.select_v_start_pos
+            end_pos = self.select_v_end_pos
+
+            # logging.debug("Drawing from %s, %s to %s. %s", start_pos[0],
+            #                                                start_pos[1],
+            #                                                end_pos[0],
+            #                                                end_pos[1] )
+
+            rect = (start_pos[0] + 0.5,
+                    start_pos[1] + 0.5,
+                    end_pos[0] - start_pos[0],
+                    end_pos[1] - start_pos[1])
+
+            # draws a light black background for the rectangle
+            ctx.set_line_width(2)
+            ctx.set_source_rgba(0, 0, 0, 0.5)
+            ctx.rectangle(*rect)
+            ctx.stroke()
+
+            # draws the dotted line
+            ctx.set_line_width(1.5)
+            ctx.set_dash([2])
+            ctx.set_line_join(cairo.LINE_JOIN_MITER)
+            ctx.set_source_rgba(*self.colour)
+            ctx.rectangle(*rect)
+            ctx.stroke()
+
+            self._debug_draw_edges(ctx)
+
+            self.position_label.pos = start_pos
+
+    def on_left_down(self, evt):
+        """ Start drag action if enabled, otherwise call super method so event will propagate """
+        if self.active.value:
+            base.SelectionMixin._on_left_down(self, evt)
+
+        base.ViewOverlay.on_left_down(self, evt)
+
+    def on_left_up(self, evt):
+        """ End drag action if enabled, otherwise call super method so event will propagate """
+        if self.active.value:
+            base.SelectionMixin._on_left_up(self, evt)
+
+        base.ViewOverlay.on_left_up(self, evt)
+
+    def on_motion(self, evt):
+        """ Process drag motion if enabled, otherwise call super method so event will propagate """
+        if self.active.value:
+            base.SelectionMixin._on_motion(self, evt)
+
+        base.ViewOverlay.on_motion(self, evt)
